feat: add matrix-testing to circle-ci config.yml
Added a lint step to run before test.
No longer relying on 'circleci node orb' because it depends on a
package-lock.json being present and newer versions do not appear to
allow overwriting this behaviour.

Fixed an invalid import from dist/ folder that was a leftover from a
previous refactoring gone wrong.

Downgraded jest versions to support testing on node8 .